+++ HISTORY.md
@@ -2,6 +2,12 @@
 
 ## Unrelease
 
+## 2.13.1
+
+1. 修订UE的FObjectFinder只能用于构造函数的问题。
+   + 增加 `UeCfg-EnableDefaultLoader` 选项控制默认的Loader是否开启
+   + 增加 `org.xresloader.ue.default_loader=EN_LOADER_MODE_DEFAULT|EN_LOADER_MODE_ENABLE|EN_LOADER_MODE_DISABLE` 选项控制单独的Message是否开启默认Loader
+
 ## 2.13.0
 
 1. 修复Excel文件过大时的报错和提示